chore: create migrations and model for products

+++ back-end/src/database/migrations/20230328201911-sales.js
@@ -12,7 +12,7 @@ const sale = (Sequelize) => ({
     references: { model: 'users', key: 'id' },
     field: 'seller_id',
   },
-  totalPrice: { type: Sequelize.DECIMAL, allowNull: false, field: 'total_price' },
+  totalPrice: { type: Sequelize.FLOAT, allowNull: false, field: 'total_price' },
   deliveryAddress: { type: Sequelize.STRING, allowNull: false, field: 'delivery_address' },
   deliveryNumber: { type: Sequelize.STRING, allowNull: false, field: 'delivery_number' },
   saleDate: { type: Sequelize.DATE, allowNull: false },
