+++ back-end/src/database/models/Sale.js
@@ -2,7 +2,7 @@ const sale = (DataTypes) => ({
   id: DataTypes.INTEGER,
   userId: { type: DataTypes.INTEGER, allowNull: false, foreignKey: true },
   sellerId: { type: DataTypes.INTEGER, allowNull: false, foreignKey: true },
-  totalPrice: { type: DataTypes.DECIMAL, allowNull: false },
+  totalPrice: { type: DataTypes.FLOAT, allowNull: false },
   deliveryAddress: { type: DataTypes.STRING, allowNull: false },
   deliveryNumber: { type: DataTypes.STRING, allowNull: false },
   saleDate: { type: DataTypes.DATE, allowNull: false },
